Nice error messages for malformed slugs
Geez, testing parsec errors is a pain in the butt. The Eq instance for
Message just checks that they're the same *constructor*; it ignores the
messages. If we actually want to test that the error message relates to
the malformed input we intended to test, we have to do this big dance
with our own equality message. Eesh.